* global resource folder implemented (all redundant resources were removed) * deferred rendering improved (removed position buffer) * position from depth calc. fixed * all illumination calcs. improved in terms of divergence control

+++ core/include/Timer.h
@@ -34,7 +34,7 @@ public:
 
 		if (QueryPerformanceCounter(&time))
 		{
-			return time.QuadPart * mSeconds;
+			return (double)time.QuadPart * mSeconds;
 		}
 
 		else
